requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
python-telegram-bot>=20.0
requests>=2.31.0
gunicorn>=21.0.0
asgiref>=3.7.0
